requirements.txt
# Data Processing
pandas>=2.0.0

# Integration & API
requests>=2.31.0
python-dotenv>=1.0.0

# Testing & Quality Assurance
pytest>=8.0.0